Update mbox exporter
Don't push to Mongo. It's way easier to keep the files in the
filesystem for examination when things go wrong.

Ignore games that have "preseason" in the subject. We've had some
Preseason tournaments.

Generate the attachment filename from the message data. This is as
close as I can get to real attachment names. It's mostly correct, but
I had to cheat with "winmail.dat" attachments.